feat(api): move admin-member route
to src
