+++ api/src/team/application/routes.js
@@ -1,3 +1,4 @@
+import { adminMemberRoutes } from './admin-member/admin-member.route.js';
 import { certificationCenterInvitationAdminRoutes } from './certification-center-invitation/certification-center-invitation.admin.route.js';
 import { certificationCenterInvitationRoutes } from './certification-center-invitation/certification-center-invitation.route.js';
 import { organizationInvitationAdminRoutes } from './organization-invitations/organization-invitation.admin.route.js';
@@ -6,6 +7,7 @@ import { prescriberInformationsRoute } from './prescriber-informations.route.js'
 
 const register = async function (server) {
   server.route([
+    ...adminMemberRoutes,
     ...certificationCenterInvitationRoutes,
     ...certificationCenterInvitationAdminRoutes,
     ...prescriberInformationsRoute,
